OHRM5X-1308: renamed and added StateContainer
