+++ installer/Controller/Installer/InstanceCreationController.php
@@ -22,6 +22,7 @@ namespace OrangeHRM\Installer\Controller\Installer;
 use OrangeHRM\Core\Vue\Component;
 use OrangeHRM\Framework\Http\Request;
 use OrangeHRM\Installer\Controller\AbstractInstallerVueController;
+use OrangeHRM\Installer\Util\StateContainer;
 
 class InstanceCreationController extends AbstractInstallerVueController
 {
@@ -30,7 +31,8 @@ class InstanceCreationController extends AbstractInstallerVueController
      */
     public function preRender(Request $request): void
     {
-        $component = new Component('instance-creation');
+        $component = new Component('instance-creation-screen');
         $this->setComponent($component);
+        StateContainer::getInstance()->setCurrentScreen(self::INSTANCE_CREATION_SCREEN);
     }
 }
