+++ installer/client/src/pages/InstanceCreationScreen.vue
